LPD-36815 Moves to correct package
