+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/item/selector/ChildSitesItemSelectorView.java
@@ -3,17 +3,17 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.publisher.web.internal;
+package com.liferay.asset.publisher.web.internal.item.selector;
 
 import com.liferay.asset.publisher.util.AssetPublisherHelper;
-import com.liferay.asset.publisher.web.internal.display.context.ParentSitesItemSelectorViewDisplayContext;
-import com.liferay.asset.publisher.web.internal.item.selector.SitesItemSelectorViewDescriptor;
+import com.liferay.asset.publisher.web.internal.display.context.ChildSitesItemSelectorViewDisplayContext;
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.ItemSelectorView;
 import com.liferay.item.selector.ItemSelectorViewDescriptorRenderer;
 import com.liferay.item.selector.criteria.GroupItemSelectorReturnType;
 import com.liferay.item.selector.criteria.group.criterion.GroupItemSelectorCriterion;
 import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -41,7 +41,7 @@ import org.osgi.service.component.annotations.Reference;
 	property = "item.selector.view.order:Integer=100",
 	service = ItemSelectorView.class
 )
-public class ParentSitesItemSelectorView
+public class ChildSitesItemSelectorView
 	implements ItemSelectorView<GroupItemSelectorCriterion> {
 
 	@Override
@@ -57,7 +57,7 @@ public class ParentSitesItemSelectorView
 	@Override
 	public String getTitle(Locale locale) {
 		return ResourceBundleUtil.getString(
-			_portal.getResourceBundle(locale), "parent-sites");
+			_portal.getResourceBundle(locale), "child-sites");
 	}
 
 	@Override
@@ -65,19 +65,24 @@ public class ParentSitesItemSelectorView
 		GroupItemSelectorCriterion groupItemSelectorCriterion,
 		ThemeDisplay themeDisplay) {
 
-		if (!groupItemSelectorCriterion.isIncludeParentSites()) {
+		if (!groupItemSelectorCriterion.isIncludeChildSites()) {
 			return false;
 		}
 
 		Group siteGroup = themeDisplay.getSiteGroup();
 
-		if (siteGroup.isLayoutPrototype() || siteGroup.isLayoutSetPrototype() ||
-			siteGroup.isRoot()) {
-
+		if (siteGroup.isLayoutPrototype() || siteGroup.isLayoutSetPrototype()) {
 			return false;
 		}
 
-		return true;
+		int groupsCount = _groupLocalService.getGroupsCount(
+			themeDisplay.getCompanyId(), siteGroup.getGroupId(), Boolean.TRUE);
+
+		if (groupsCount > 0) {
+			return true;
+		}
+
+		return false;
 	}
 
 	@Override
@@ -87,9 +92,9 @@ public class ParentSitesItemSelectorView
 			PortletURL portletURL, String itemSelectedEventName, boolean search)
 		throws IOException, ServletException {
 
-		ParentSitesItemSelectorViewDisplayContext
-			parentSitesItemSelectorViewDisplayContext =
-				new ParentSitesItemSelectorViewDisplayContext(
+		ChildSitesItemSelectorViewDisplayContext
+			childSitesItemSelectorViewDisplayContext =
+				new ChildSitesItemSelectorViewDisplayContext(
 					groupItemSelectorCriterion,
 					(HttpServletRequest)servletRequest, _assetPublisherHelper,
 					portletURL);
@@ -99,7 +104,7 @@ public class ParentSitesItemSelectorView
 			portletURL, itemSelectedEventName, search,
 			new SitesItemSelectorViewDescriptor(
 				(HttpServletRequest)servletRequest,
-				parentSitesItemSelectorViewDisplayContext));
+				childSitesItemSelectorViewDisplayContext));
 	}
 
 	private static final List<ItemSelectorReturnType>
@@ -108,6 +113,9 @@ public class ParentSitesItemSelectorView
 
 	@Reference
 	private AssetPublisherHelper _assetPublisherHelper;
+
+	@Reference
+	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private ItemSelectorViewDescriptorRenderer<GroupItemSelectorCriterion>
